+++ src/ndata.rs
@@ -29,12 +29,12 @@ use crate::dataarray::*;
 use crate::databytes::*;
 
 /// Initialize global storage of data. Call only once at startup.
-pub fn init() -> (((usize,usize),(usize,usize)),((usize,usize),(usize,usize)),((usize,usize),(usize,usize))) {
+pub fn init() -> (((u64,u64),(u64,u64)),((u64,u64),(u64,u64)),((u64,u64),(u64,u64))) {
   (DataObject::init(), DataArray::init(), DataBytes::init())
 }
 
 /// Mirror global storage of data from another process. Call only once at startup.
-pub fn mirror(data_ref:(((usize,usize),(usize,usize)),((usize,usize),(usize,usize)),((usize,usize),(usize,usize)))) {
+pub fn mirror(data_ref:(((u64,u64),(u64,u64)),((u64,u64),(u64,u64)),((u64,u64),(u64,u64)))) {
   DataObject::mirror(data_ref.0.0, data_ref.0.1);
   DataArray::mirror(data_ref.1.0, data_ref.1.1);
   DataBytes::mirror(data_ref.2.0, data_ref.2.1);
